+++ gf16/gf16_shuffle128_sve_common.h
@@ -16,7 +16,7 @@ static HEDLEY_ALWAYS_INLINE void gf16_shuffle128_sve_calc_tables(const void *HED
 	UNUSED(scratch);
 	#endif
 	
-	svint16_t val1 = svld1_s16(svwhilelt_b16(0u, srcCount), (int16_t*)coefficients);
+	svint16_t val1 = svld1_s16(svwhilelt_b16((uint32_t)0, (uint32_t)srcCount), (int16_t*)coefficients);
 	svint16_t val2 = gf16_vec_mul2_sve(val1);
 	svint16_t val4 = gf16_vec_mul2_sve(val2);
 	svint16_t val8 = gf16_vec_mul2_sve(val4);
