+++ gf16/gf16_shuffle512_sve2.c
@@ -50,7 +50,7 @@ static HEDLEY_ALWAYS_INLINE void gf16_shuffle512_sve2_calc_tables(
 		svld1_u8(svwhilelt_b8(0, 64), (uint8_t*)scratch + 64)
 	);
 	
-	svint16_t val = svld1_s16(svwhilelt_b16(0u, srcCount), (int16_t*)coefficients);
+	svint16_t val = svld1_s16(svwhilelt_b16((uint32_t)0, (uint32_t)srcCount), (int16_t*)coefficients);
 	// dupe 16b elements across 128b vector
 	val = svtbl_s16(val, NOMASK(svlsr_n_u16, svindex_u16(0, 1), 3));
 	// (alternative idea to do a 16->64b extended load, unpack, mul-lane)
